refactoring createProductMVC
