+++ src/main/java/pl/pozsda19/electroshop/service/ProductService.java
@@ -6,8 +6,10 @@ import pl.pozsda19.electroshop.domain.Group;
 import pl.pozsda19.electroshop.domain.Product;
 import pl.pozsda19.electroshop.domain.Subcategory;
 import pl.pozsda19.electroshop.domain.dto.ProductEntityReading;
+import pl.pozsda19.electroshop.domain.dto.ProductEntityWriting;
 import pl.pozsda19.electroshop.domain.dto.ProductMapper;
 import pl.pozsda19.electroshop.exception.DuplicateProductCodeException;
+import pl.pozsda19.electroshop.exception.ProductNotFoundException;
 import pl.pozsda19.electroshop.repository.ProductRepository;
 
 import java.util.List;
@@ -26,6 +28,15 @@ public class ProductService {
         this.productMapper = productMapper;
     }
 
+    public Product createProduct(ProductEntityWriting toProductEntity){
+        Product product = productMapper.writeProductEntity(toProductEntity);
+        productRepository.save(product);
+        return product;
+    }
+    public boolean productExists(String code){
+        return productRepository.existsByCode(code);
+    }
+
     public Set<ProductEntityReading> showAllProducts(){
         return productRepository.findAll()
                 .stream()
@@ -36,9 +47,11 @@ public class ProductService {
     public Optional<Product> retrieveProductByCode(String code){
         return productRepository.findByCode(code);
     }
-    public Optional<ProductEntityReading> showProductByCode(String code) {
+
+    public Optional<ProductEntityReading> getProductByCode(String code) {
         return productRepository.findByCode(code).map(product -> productMapper.readProductEntity(product));
     }
+
     public Set<ProductEntityReading> retrieveProductsByCategory(Category category){
         return productRepository.findProductsByCategory(category).stream()
                 .map(product -> productMapper.readProductEntity(product))
@@ -56,7 +69,7 @@ public class ProductService {
                 .collect(Collectors.toSet());
     }
 
-    public String addProduct(Product product){
+   public String addProduct(Product product){
         productRepository.findByCode(product.getCode())
                 .ifPresent(product1 -> throwDuplicateProductCodeException(product.getCode()));
         productRepository.save(product);
@@ -68,5 +81,17 @@ public class ProductService {
 
     public String addProducts(List<Product> products) {
         return null;
+    }
+    public Product updateProduct(ProductEntityWriting productEntityWriting){
+        Optional<Product> productFromDB = productRepository.findByCode(productEntityWriting.getCode());
+        if(productFromDB.isEmpty()) {
+            throw new ProductNotFoundException("Brak produktu o podanym kodzie");
+        }
+
+        productRepository.delete(productFromDB.get());
+        Product productUpdated = productMapper.writeProductEntity(productEntityWriting);
+        productRepository.save(productUpdated);
+        return productUpdated;
+
     }
 }
