+++ src/main/resources/data.sql
@@ -1,8 +1,8 @@
-insert  into PRODUCT(id, code, name, category, subcategory,groupo, imageURL, description, price, quantity)values (nextval('product_Seq'),'330435', 'Poziomica cyfrowa BOSCH DNM 60L', 'TECHNIKAPOMIAROWA', 'POZIOMICE','', 'http://www.elektroserw.pl/zdjecia/47_1.jpg', 'description', 850.00, 1);
-insert  into PRODUCT(id, code, name, category, subcategory,groupo, imageURL, description, price, quantity)values (nextval('product_Seq'),'345562', 'Poziomica cyfrowa BOSCH DNM 120L', 'TECHNIKAPOMIAROWA', 'POZIOMICE','', 'http://www.elektroserw.pl/zdjecia/46_1.jpg', 'description', 625.00, 1);
+insert  into PRODUCT(id, code, name, category, subcategory, imageURL, description, price, quantity)values (nextval('product_Seq'),'330435', 'Poziomica cyfrowa BOSCH DNM 60L', 'TECHNIKAPOMIAROWA', 'POZIOMICE', 'http://www.elektroserw.pl/zdjecia/47_1.jpg', 'description', 850.00, 1);
+insert  into PRODUCT(id, code, name, category, subcategory, imageURL, description, price, quantity)values (nextval('product_Seq'),'345562', 'Poziomica cyfrowa BOSCH DNM 120L', 'TECHNIKAPOMIAROWA', 'POZIOMICE', 'http://www.elektroserw.pl/zdjecia/46_1.jpg', 'description', 625.00, 1);
 
-insert  into PRODUCT(id, code, name, category, subcategory,groupo, imageURL, description, price, quantity)values (nextval('product_Seq'),'345563', 'Wykrywacz przewodów BOSCH GMS 120', 'TECHNIKAPOMIAROWA', 'WYKRYWACZE','', 'http://www.elektroserw.pl/zdjecia/45_1.jpg', 'description', 500.00, 1);
-insert  into PRODUCT(id, code, name, category, subcategory,groupo, imageURL, description, price, quantity)values (nextval('product_Seq'),'0601010000', 'Wykrywacz przewodów BOSCH DMF 10 Zoom', 'TECHNIKAPOMIAROWA', 'WYKRYWACZE','', 'http://www.elektroserw.pl/zdjecia/29_2.jpg', 'description', 500.00, 1);
+insert  into PRODUCT(id, code, name, category, subcategory, imageURL, description, price, quantity)values (nextval('product_Seq'),'345563', 'Wykrywacz przewodów BOSCH GMS 120', 'TECHNIKAPOMIAROWA', 'WYKRYWACZE', 'http://www.elektroserw.pl/zdjecia/45_1.jpg', 'description', 500.00, 1);
+insert  into PRODUCT(id, code, name, category, subcategory, imageURL, description, price, quantity)values (nextval('product_Seq'),'0601010000', 'Wykrywacz przewodów BOSCH DMF 10 Zoom', 'TECHNIKAPOMIAROWA', 'WYKRYWACZE', 'http://www.elektroserw.pl/zdjecia/29_2.jpg', 'description', 500.00, 1);
 
 insert  into PRODUCT(id, code, name, category, subcategory, groupo, imageURL, description, price, quantity)values (nextval('product_Seq'),'453', 'Wykrywacz przewodów BOSCH DMF 10 Zoom', 'CZESCIZAMIENNE', 'UZYWANE', 'WIRNIKI', 'http://www.elektroserw.pl/zdjecia/29_2.jpg', 'description', 500.00, 1);
 insert  into PRODUCT(id, code, name, category, subcategory, groupo, imageURL, description, price, quantity)values (nextval('product_Seq'),'111', 'Wykrywacz przewodów BOSCH DMF 10 Zoom', 'CZESCIZAMIENNE', 'UZYWANE', 'OBUDOWY', 'http://www.elektroserw.pl/zdjecia/29_2.jpg', 'description', 500.00, 1);
