+++ src/main/resources/templates/productCreated.html
@@ -10,7 +10,7 @@
 <p th:text="${message}" ></p>
 
 <a th:href="@{/showProducts}">Lista Produktów</a><br>
-<a th:href="@{/productAdministration/addProduct}">Dodaj Produkt</a>
+<a th:href="@{/showProducts/addProduct}">Dodaj Produkt</a>
 
 </body>
 </html>
