+++ src/test/java/pl/pozsda19/electroshop/service/ProductServiceTest.java
@@ -11,6 +11,7 @@ import pl.pozsda19.electroshop.domain.Group;
 import pl.pozsda19.electroshop.domain.Product;
 import pl.pozsda19.electroshop.domain.Subcategory;
 import pl.pozsda19.electroshop.domain.dto.ProductEntityReading;
+import pl.pozsda19.electroshop.domain.dto.ProductEntityWriting;
 import pl.pozsda19.electroshop.exception.DuplicateProductCodeException;
 
 import java.math.BigDecimal;
@@ -26,6 +27,29 @@ class ProductServiceTest {
 
     @Autowired
     private ProductService productService;
+
+    @Test
+    void createProductShouldAddANewProductToTheDatabase(){
+        //given
+        ProductEntityWriting product = new ProductEntityWriting("1234", "Wyrzynarka", "ELEKTRONARZEDZIA", "AKUMULATOROWE"
+                , "WYRZYNARKI", "image","wyrzyna dobrze", 30.00,30);
+        String expectedCode = "1234";
+        //when
+        String actualProductCode = productService.createProduct(product).getCode();
+        //then
+        Assert.assertEquals(expectedCode, actualProductCode);
+
+    }
+
+    @Test
+    void showAllProductsEndpointShouldReturn7Products(){
+        //given
+        int expectedSize = 7;
+        //when
+        int actualSize = productService.showAllProducts().size();
+        //then
+        Assert.assertEquals(expectedSize,actualSize);
+    }
 
     @Test
     void showProductsByCategoryShouldReturnSetSize3ForCZESCIZAMIENNE(){
@@ -95,24 +119,22 @@ class ProductServiceTest {
         String code = "0601010000";
         String expectedCategory = "TECHNIKAPOMIAROWA";
         //when
-        Optional<ProductEntityReading> readProductModel = productService.showProductByCode(code);
+        Optional<ProductEntityReading> readProductModel = productService.getProductByCode(code);
         String actualCategory = readProductModel.get().category;
         //then
         Assert.assertEquals(expectedCategory,actualCategory);
     }
 
-
-
     @Test()
-    void addProductShouldThrowDuplicateCodeExceptionForProductWithCode0601010000() {
+    void createProductShouldThrowDuplicateCodeExceptionForProductWithCode0601010000() {
        //given
-        Product product = new Product("0601010000", "bla", Category.TECHNIKAPOMIAROWA, Subcategory.WYKRYWACZE
-                ,"url","des", BigDecimal.valueOf(500));
+        ProductEntityWriting product = new ProductEntityWriting("0601010000", "bla", "TECHNIKAPOMIAROWA", "WYKRYWACZE",
+                "ORINGI","url","des", 500.00, 1);
         String expectedMessage = "Product code 0601010000 in use";
 
         //when
         DuplicateProductCodeException exception = assertThrows(DuplicateProductCodeException.class, ()->
-                productService.addProduct(product));
+                productService.createProduct(product));
         String actualMessage = exception.getMessage();
 
         //then
@@ -132,5 +154,24 @@ class ProductServiceTest {
         //then
         Assert.assertEquals(expectedCategoryString,actualCategoryString);
     }
+
+    @Test
+    void updateProductWithKnownIdShouldUpdatedTheProduct(){
+        //given
+        ProductEntityWriting product = new ProductEntityWriting("453"
+                , "Wykrywacz przewodów BOSCH DMF 10 Zoom", "CZESCIZAMIENNE"
+                , "UZYWANE", "WIRNIKI"
+                , "http://www.elektroserw.pl/zdjecia/29_2.jpg", "opis"
+                , 500.00, 1);
+        String expectedDescription = "opis";
+        //when
+        productService.updateProduct(product);
+        String actualDescription = productService.retrieveProductByCode("453").get().getDescription();
+        //then
+        Assert.assertEquals(expectedDescription,actualDescription);
+
+    }
+
+
 
 }
